move test helpers into separate modules, two FakeIOs to handle different outputs for "gets"
