+++ test/support/ui_helper.exs
@@ -1,0 +1,6 @@
+defmodule UiTestHelper do
+  def retry_after_error(io) do
+    io.puts("something went wrong")
+    :value_after_error
+  end
+end
